Include Debian package version in the tool (#437)
When running the cjxl/djxl tools we include two versions, the library
semantic version numbers (MAJOR.MINOR.PATCH) as reported by the library
and an arbitrary string that comes from the package. The arbitrary
string is normally the git hash or other version or build information.
For the Debian packages this now includes the Debian package version
string.