CMake setting to bundle libskcms.a into libjxl.a (#441)
skcms is only a single .cc file with a C API, but not widely available
as a standalone package or library. We currently require to use a CMS in
the encoder side of the library to convert from arbitrary color spaces
when encoding, among probably other things. The user can select from
either skcms (the default) or lcms2 for the CMS library. In the skcms
case libjxl.so was bundling the libskcms.a symbols (since skcms is built
as a static library) but libjxl.a wasn't, resulting in an unusable
static library being packaged and installed.

This patch adds a default-on setting to bundle skcms.cc into libjxl.a
when using skcms, however, we rename all the function from `skcms_` to
`jxl_skcms_` to avoid issues with potential projects that would
include skcms from elsewhere. Until there's a packaged version of skcms
or we provide an external way to provide a CMS on the encoder side this
at least makes libjxl.a link a sample program.

Fixes #428